+++ tobaccocessation/settings_shared.py
@@ -22,7 +22,6 @@ TEMPLATES[0]['OPTIONS']['context_processors'].append(  # noqa
 
 INSTALLED_APPS += [  # noqa
     'sorl.thumbnail',
-    'typogrify',
     'bootstrapform',
     'django_extensions',
     'pagetree',
