+++ tobaccocessation/urls.py
@@ -1,18 +1,20 @@
-import django.contrib.auth.views
-import djangowind.views
-import os.path
-
 from django.conf import settings
 from django.conf.urls import include, url
 from django.contrib import admin
-from django.contrib.auth import views as auth_views
+from django.contrib.auth.views import password_change, password_change_done, \
+    password_reset, password_reset_done, password_reset_complete, \
+    password_reset_confirm
+import django.contrib.auth.views
 from django.views.generic import TemplateView
+from django.views.static import serve
+import djangowind.views
 from registration.backends.default.views import RegistrationView
 
-from tobaccocessation.main.models import CreateAccountForm
+from tobaccocessation.main.forms import CreateAccountForm
 from tobaccocessation.main.views import (
     index, edit_page, page, create_profile,
 )
+
 
 admin.autodiscover()
 
@@ -35,37 +37,35 @@ if hasattr(settings, 'CAS_BASE'):
                             'djangowind.views.logout',
                             {'next_page': redirect_after_logout})
 
+
 urlpatterns = [
     logout_page,
     admin_logout_page,
     auth_urls,
-    url(r'^crossdomain.xml$', django.views.static.serve,
-        {'document_root': os.path.abspath(os.path.dirname(__file__)),
-         'path': 'crossdomain.xml'}),
     url(r'^about/', TemplateView.as_view(
         template_name="flatpages/about.html")),
-    url(r'^help/', TemplateView.as_view(template_name="flatpages/help.html")),
+    url(r'^help/', TemplateView.as_view(
+        template_name="flatpages/help.html")),
     url(r'^contact/', TemplateView.as_view(
         template_name="flatpages/contact.html")),
     url(r'^$', index, name="index"),
 
-    url(r'^accounts/register/$', RegistrationView.as_view(
-        form_class=CreateAccountForm), name='registration_register'),
-
-    # override the default urls for pasword
-    url(r'^password/change/$', auth_views.password_change,
+    # override the default urls for password
+    url(r'^password/change/$', password_change,
         name='password_change'),
-    url(r'^password/change/done/$', auth_views.password_change_done,
+    url(r'^password/change/done/$', password_change_done,
         name='password_change_done'),
-    url(r'^password/reset/$', auth_views.password_reset,
+    url(r'^password/reset/$', password_reset,
         name='password_reset'),
-    url(r'^password/reset/done/$', auth_views.password_reset_done,
+    url(r'^password/reset/done/$', password_reset_done,
         name='password_reset_done'),
-    url(r'^password/reset/complete/$', auth_views.password_reset_complete,
+    url(r'^password/reset/complete/$', password_reset_complete,
         name='password_reset_complete'),
     url(r'^password/reset/confirm/(?P<uidb64>[0-9A-Za-z]+)-(?P<token>.+)/$',
-        auth_views.password_reset_confirm, name='password_reset_confirm'),
+        password_reset_confirm, name='password_reset_confirm'),
 
+    url(r'^accounts/register/$', RegistrationView.as_view(
+        form_class=CreateAccountForm), name='registration_register'),
     url(r'^accounts/', include('registration.backends.default.urls')),
 
     url(r'^admin/', include(admin.site.urls)),
@@ -79,8 +79,8 @@ urlpatterns = [
     url(r'^activity/quiz/', include('quizblock.urls')),
     url(r'^quizblock/', include('quizblock.urls')),
 
-    url(r'^uploads/(?P<path>.*)$', django.views.static.serve,
-        {'document_root': settings.MEDIA_ROOT}),
+    url(r'^uploads/(?P<path>.*)$',
+        serve, {'document_root': settings.MEDIA_ROOT}),
 
     url(r'^pagetree/', include('pagetree.urls')),
